feature: bugfixes & hotkeys

+++ src/main/java/org/taurus/aya/server/GroupRepository.java
@@ -8,6 +8,5 @@ import org.taurus.aya.server.entity.Group;
 import java.util.List;
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Integer > {
-
+public interface GroupRepository extends JpaRepository<Group, Long > {
 }
